toolchain: add DSU loader script and have options to skip repo sync

+++ nuke.local.sh
@@ -1,3 +1,10 @@
 #!/bin/bash
 
-repo forall -vc "git reset --hard ; git clean -fdx"
+
+read -p "Are you sure? THIS WILL NUKE ALL LOCAL CHANGES!: " -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    repo forall -vc "git reset --hard ; git clean -fdx"
+fi
+
+
